Parse datetime utils: handle shorter months
Adding a "month" timedelta is not that easy! We'd like to have used
dateutil.relativedelta here, but we don't have dateutil in here, and
I don't want to pull it in for just this.

Just doing .replace(month=month+delta) isn't enough, because for
some values that's a ValueError: eg. you can't add 1month to
March 31st this way, because that would end up being April 31st, and
that doesn't exist.

Instead, .replace the month for a datetime with day=1
(so Mar31 -> Apr01) and then add the original number of days as a
timedelta, which does carry over the month
(so Apr01 + 30 days -> Apr31 -> May01).