rewrite mock_storage usage in TestLabels
same as the previous commit, but for TestGetLabel.
I note that these TestGetLabel tests also use a capability, but now
they get their own capability defined. And so, now each test class
has its own setup, as opposed to everything using the same capabilities
always.
All in all, the point really is to make the asserted-on value be closer
to the test itself.