Make sure tests are not using remote HTTP calls
Calling out to the internet in the tests is strictly forbidden.

Instead, make a fake `http://` address that is intercepted by the
resolver's rules, and use that. (I think we can't use a file directly,
because all the blueprints are just strings, not files)

Since we don't want to call out to the internet to get types.yaml,
make a minimal quasi-types.yaml locally, similar to what we did for
cloudify/tests/. (this one happens to be smaller)

Obviously, this makes the tests WAY faster, and they are now decently
fast, too: before this, the whole suite took on the order of 45 seconds
on my machine (~960 tests). With this, and with xdist=4, I can get it
to go as low as 3.5 seconds.