RD-1165 Add get_group_capability (#752)
* rewrite mock_storage usage in TestGetCapability

Instead of having static capabilities for all tests ever, just
define them next to the tests that use them. Saves a bit of my sanity
when looking at the MockStorage class.

* rewrite mock_storage usage in TestLabels

same as the previous commit, but for TestGetLabel.
I note that these TestGetLabel tests also use a capability, but now
they get their own capability defined. And so, now each test class
has its own setup, as opposed to everything using the same capabilities
always.
All in all, the point really is to make the asserted-on value be closer
to the test itself.

* Make sure tests are not using remote HTTP calls

Calling out to the internet in the tests is strictly forbidden.

Instead, make a fake `http://` address that is intercepted by the
resolver's rules, and use that. (I think we can't use a file directly,
because all the blueprints are just strings, not files)

Since we don't want to call out to the internet to get types.yaml,
make a minimal quasi-types.yaml locally, similar to what we did for
cloudify/tests/. (this one happens to be smaller)

Obviously, this makes the tests WAY faster, and they are now decently
fast, too: before this, the whole suite took on the order of 45 seconds
on my machine (~960 tests). With this, and with xdist=4, I can get it
to go as low as 3.5 seconds.

* Rewrite get_secret usage in tests

Some of the tests don't even need get_secret to be pased at all!
(which makes sense, since get_secret is a runtime func, and a lot of
the tests only do parsing, not anything runtime)

The point is:
1. Don't require `self._mock_storage` to be set. The function now
  ONLY returns, as opposed to both setting the attribute AND returning,
  which was ugly
2. The method isn't really needed at all. Replace it by just calling
  the class directly, but I'll still keep a handy shortcut on the
  test class, but as just a property instead.

* drop unused method

i think it might have been used at some point??? it's not right now

* GetCapability: drop a range+len

Use an enumerate instead. Makes for nicer code. I'm going to copy this
class to base GetGroupCapa out of it, so I want to make it nice first.

* Simplify InterDeploymentDependencyCreatingFunction

I have no idea why was this written like that.
It did absolutely nothing.

* Implement GetGroupCapability

The class itself, and the tests, are really boring and trivial.
The interesting logic is in the storage side, ie. in cloudify-manager.
Other than that, it's mostly a copy of GetCapability.