RD-2579 _SetNodeInstanceStateTask: force-set the state
We used to first fetch the instance, to figure out the current
version, and then set the new state anyway.
So, the only reason we fetched it, was so that our update always
succeeds - we didn't react to the received state in any way.
We can simply just cut out the middleman, and force-set the
state directly, without fetching it.

And anyway, force-setting, like we (essentially) did before,
and do now, does make sense for the state - instance state is never
changed concurrently, only runtime-properties are (and that's what
version is for).