Drop CallbackRequestResponseHandler
I'm super happy to simplify this one! `_RequestResponseHandlerBase`
used to have two subclasses, the callback one and the blocking
one. I've never been able to hold in my head how do these subclasses
actually work. So now it's simpler.

Well, the callback one is only used in one place, and I'm about to
refactor that place, so it can just go.
The blocking one is used in several places, so it must stay here, but
now it can be merged with its superclass.