implement a _WorkflowTaskHandler
I've just removed the callback handler, and said I'm gonna use
something else instead, well this is the something else. It's a
simplified version of that handler.

This handler binds to a single queue named "response_{exec-id}",
and sends tasks with a reply-to, to this queue.

This means all tasks inside of a single execution will send their
responses to this single queue, as opposed to the previous idea
of having a response-queue per task.

The task responses are still identified by their correlation_id,
and callbacks are run based on that.

Also, this adds a "buffer", to store task responses received
before we actually start waiting for them (ie. before we add a
callback). This happens when we resume an execution and some
responses are already waiting for us.