Refactor _TaskDispatcher
Instead of starting an amqp client per task, the dispatcher now
starts a single client per vhost used, and attaches that new handler
that we just implemented, to it.

Since we don't need to carry client+handler around per task, the
make_subtask stuff is unnecessary, because now both send_task
and wait_for_result are able to work in terms of just the task.

This means we no longer have the idea of a "workflow_task" (the
Task instance) and a "task" (the weird dict returned from
make_subtask), we just work in terms of a "task" (ie. the Task
instance).