Cleanup when the execution finishes
Now that we have a response-queue-per-execution, we must remove it
when the execution finishes.
I'm making it proper methods here, possibly we can also clean up
something else??? Is there something we should? I don't think so right
now?

But don't do it if the execution is cancelled, because for a cancelled
execution, maybe the agents are still running operations, and they
want to send back responses, which we'll need to pick up on a resume.
This is similar to what we've done before, but now it's just 1 queue
potentially-leaking, rather than 1 queue per still-running task.