Move the task callback from dispatcher into the handler
The callback is always the same function, so instead of working
in terms of callback functions, we can work in terms of storing
tasks, and always call the same callback function with that stored
task.

Also, no need to pass through .async_result, it is just available
on the task.