RD-513 Use a single AMQP queue per execution (#798)
* Remove pointless log

I hate this log, it keeps spamming mgmtworker.log, it's never been
useful, and it assumes that a handler has a .routing_key; I'm going
to break that assumption soon

* Drop CallbackRequestResponseHandler

I'm super happy to simplify this one! `_RequestResponseHandlerBase`
used to have two subclasses, the callback one and the blocking
one. I've never been able to hold in my head how do these subclasses
actually work. So now it's simpler.

Well, the callback one is only used in one place, and I'm about to
refactor that place, so it can just go.
The blocking one is used in several places, so it must stay here, but
now it can be merged with its superclass.

* implement a _WorkflowTaskHandler

I've just removed the callback handler, and said I'm gonna use
something else instead, well this is the something else. It's a
simplified version of that handler.

This handler binds to a single queue named "response_{exec-id}",
and sends tasks with a reply-to, to this queue.

This means all tasks inside of a single execution will send their
responses to this single queue, as opposed to the previous idea
of having a response-queue per task.

The task responses are still identified by their correlation_id,
and callbacks are run based on that.

Also, this adds a "buffer", to store task responses received
before we actually start waiting for them (ie. before we add a
callback). This happens when we resume an execution and some
responses are already waiting for us.

* Refactor _TaskDispatcher

Instead of starting an amqp client per task, the dispatcher now
starts a single client per vhost used, and attaches that new handler
that we just implemented, to it.

Since we don't need to carry client+handler around per task, the
make_subtask stuff is unnecessary, because now both send_task
and wait_for_result are able to work in terms of just the task.

This means we no longer have the idea of a "workflow_task" (the
Task instance) and a "task" (the weird dict returned from
make_subtask), we just work in terms of a "task" (ie. the Task
instance).

* Use the new TaskDispatcher interface

We changed send_task and wait_for_result, and dropped make_subtask.
So the other parts of the system can now use the new convention.

Also, just *a**kw the methods on the WorkflowContextHandler, because
well, they're just a proxy/facade to the real thing.

* Cleanup when the execution finishes

Now that we have a response-queue-per-execution, we must remove it
when the execution finishes.
I'm making it proper methods here, possibly we can also clean up
something else??? Is there something we should? I don't think so right
now?

But don't do it if the execution is cancelled, because for a cancelled
execution, maybe the agents are still running operations, and they
want to send back responses, which we'll need to pick up on a resume.
This is similar to what we've done before, but now it's just 1 queue
potentially-leaking, rather than 1 queue per still-running task.

* Move the task callback from dispatcher into the handler

The callback is always the same function, so instead of working
in terms of callback functions, we can work in terms of storing
tasks, and always call the same callback function with that stored
task.

Also, no need to pass through .async_result, it is just available
on the task.

* flake

* fix test_workflow_update_execution_status_set_to_false

just add a mock. I'm gonna rewrite all of these tests eventually.
But not now.