RD-2618 Move the subprocess bits out of dispatch.py (#801)
* Expose a few more attributes on the ctx

we use them by accessing the raw context dict, it'd be nicer if we
exposed them properly instead

* Remove all the subprocess-related things from dispatch.py

I'll move it somewhere else!

* drop test_dispatch_resume

what does it have to do with resuming? well, nothing. Anyway, this
method has left the building.

* remove the other dispatch-to-subprocess tests

that needs to be tested inside cloudify-agent

* simplify test_parse_utc_datetime_no_date

i mean, otherwise, it fails for me between midnight and 2am
(and depending on the timezone it'll fail in different times)

this was because `utcnow().replace(hour=...)` is still.. "yesterday"
in utc. Anyway, just asserting on the hour is kinda the point.

* remove now-disused things from test_dispatch