task set_state: send an event only if it's not stored
we still keep compat with non-stored tasks, ie. tasks that are not
using a tasks graph. For those, we construct the event and send it
separately. Those aren't stored, so there's no server-side "operation
update" to talk about.

For tasks that are stored, we just send the operation update request.
The event will be inserted by the server.