RD-515 Don't create events for operations (#805)
* Factor out formatting the event message

Simple enough, just factor out those two functions, figuring out
the message, and the event type.

I'm gonna use them in c-manager.

* Add result, exception & causes to the operation-update request

Now also sending in those values when updating an operation.
This is so that the server can insert an event automatically.

In the future, maybe we'll also start storing the result as well.

* Task on failure: dont send the event

No need. It will be automatically inserted in updating the state.
(in the next commits)

Also it'll happen now for all tasks, also subgraphs. This change
doesn't matter. We won't be inserting events (yet?) for subgraphs
anyway. But now it's up to the server.

* task set_state: send an event only if it's not stored

we still keep compat with non-stored tasks, ie. tasks that are not
using a tasks graph. For those, we construct the event and send it
separately. Those aren't stored, so there's no server-side "operation
update" to talk about.

For tasks that are stored, we just send the operation update request.
The event will be inserted by the server.

* workflow-ctx: leave it for the task to send events or not

no need to send events from here now. The task will do it (if it's
not stored), or the server will (if the task is stored). Anyway,
this part now only sets the state, but now passing the
result/exception through as well.

* drop this event

it's kinda pointless. We insert an invent for the TASK_SENT state
change now though.

* drop the events from dry-run tasks

those are tbh entirely pointless. Dry-run does no sending to speak of.
Just update the state and we're done here.

* switch out local-task event sending a bit

The idea is:
1. before the task is queued, we set TASK_SENT (which is not stored
   on the server, due to the if in _update_stored_state, so it's
   essentially a no-op)
2. when the task is starting, we set TASK_STARTED, with the same
   semantics as 1
3. when the task is done, we set TASK_SUCCEEDED, and that _is_ stored.
   The server will not insert an event anyway, because it doesn't do
   it for local tasks.

There's two kinds of important local tasks:
- in regular workflows, things like SendEvent and SetNodeInstanceState
  are local tasks, but those are somewhat built-in/internal, and we
  don't want to store events for those anyway, and never did.
- in `cfy local` workflows, your regular plugin/script tasks are
  local tasks, and we presumably could use events for these, why not.

So for now I want to keep this, but in the future, after RD-2650
does away with the first kind of local task, then I'll remove the
`if state` from _update_stored_task, and all state changes will be
recorded for the local tasks (who then will be only the second kind)

* Update tests for the local task events

The change in the previous commit did effectively drop one local task
event.
So update the tests to expect one less.

* fix mocks in this test

tasks in this test cant just use None as the workflow context,
because now we touch it in set_state. Pass a mock instead.

* flake

* local task: set the state before setting a result

otherwise, the set-state might not happen before the workflow is done

* rewrite the event message helpers using dicts