RD-518 Insert logs/events via a REST call (#806)
* _publish_message: send in logs/events via REST rather than AMQP

Just use a rest-client, and boom done.

* stop using AMQPWrappedThread

we now don't need the amqp client, so no need to use a thread class
which creates one.

* no need to prepare the events publisher

* drop a now-unused module

wooo!

* Update execution state AFTER sending in the event.

It's now synchronous! This means that when the execution-state is
changed, all events are already stored. No need for the 3-second
wait anymore.

* flake

* drop unnecessary mocks

the module is gone, no need to mock it