RD-2650 Stop using LocalTask for the builtin tasks
Instead of using LocalWorkflowTask for all the builtin tasks
(such as set-node-instance-state, send-event, etc), make them
each be a _subclass_ of WorkflowTask.

That means they're no longer tied to LocalWorkflowTask, which is
for sending actual tasks (in cfy local) as well.
Specifically, we'll be able to just skip sending the SENDING/SENT
state updates, because they're not needed for these builtin tasks.