RD-519 RemoteContextHandler: use a stored rest-client (#814)
* move two classes around a bit

this is so that RemoteContextHandler is right next to its abstract
base class, not separate by the taskdispatcher stuff

* RemoteContextHandler: use a stored rest-client

instead of calling get_rest_client in every method, store one on the
instance

* WorkflowContext: factor out getting nodes/instances

the workflow context already has a Handler to it, either local or
remote, and THAT is the point of indirection. The `if local` in
workflow context itself, needs to go.

* set .blueprint/.deployment up front

no need to put those under the `with current:` push. The less
initialization we can do under a context, the better