set-node-instance-state task: handle non-stored tasks
The automatic execution on state update, only happens if we even
update the operation state, which happens only in cases where
the tasks are in a graph.

If the tasks aren't in a graph, they're not stored and not updated,
and there needs to be an explicit implementation of the task.
Unfortunate, but still possible to do.

This fixes some inte-tests, eg `integration_tests/tests/agentless_tests/test_executions.py::ExecutionsTest::test_cancel_execution `