RD-2679 Clear the queue-bind cache when deleting agents (#827)
See the docstring of `_task_deletes_exchange` for explanation.

As always, cache invalidation is one of the Hard™ things!

We clear the whole cache because that is trivially correct: otherwise,
it'd be hard to figure out what exchange should we even remove from
the cache, because the delete task is a mgmtworker task (and so,
target = `cloudify.management`), on the `/` (not tenant) vhost - and
so, separate Handler instances. And the exchange/queue names aren't
necessarily the same as the node-instance name, because of proxied
agents.