RD-2751 Attach the execution id to the events/logs create
Turns out it's not always the case that we only create events/logs
via a request authenticated by an execution token - when running hooks,
we are authenticated by the user's rest token instead.

So, allow passing in the execution id, so that requests authenticated
by a rest token, know which execution are they related to.