RD-2802 Add `hidden` input property (#840)
* Add `hidden` input property to DSL

The property required at least cloudify_dsl_1_3.

An input which is both hidden and required must also have a default
value set.

* Don't force `hidden` property to be always present

It is not necessary to always store the `hidden` flag (just as it is not
necessary to always store `required` flag) of an input.  In case the
flag is not set in the blueprint, just don't store it at all.