Use order-ignoring comparisons for several namespace things
Apparently, namespaces return those things in no ordering in
particular. That used to be previously asserted via the
order-ignoring assertItemsEqual, but now that we've changed it
to the order-sensitive assertEqual, it must be set()-ed or similar.

Anyway, I find some hilarity in the fact that we specifically
ordered the expected items, eg `[workflows[1], workflows[0]]`,
AND THEN ignored the order, by way of assertItemsEqual.

Which suggests that perhaps namespaces are fucked. But I'm not sure
that ordering is even part of its contract at all.