RD-3115 dsl-parser tests: stop using testtools (#873)
* Mark test_merging_namespaces_mapping xfail

This is failing because we used to (via assertItemsEqual) only really
assert on the keys. Now we do assert on the values as well, and turns
out this test was wrong all along. Is the behaviour wrong as well?
Who knows! That's up to RD-3097 to figure out.

* Use the `with` form of assertRaises

More modern this way. Also, non-testtools assertRaises doesn't
return the exception otherwise.

* s/assertItemsEqual/assertEqual/

There's no point in using assertItemsEqual in these tests, where
we only care about equality indeed. ItemsEqual was the 2.7-only way
of saying "equal, but without checking for ordering".

For added hilarity, some tests DID care about ordering (note: that
one test specifically puts [1] first and [0] later), and then used
the order-ignoring assertion. Go figure.

* test_inputs: s/assertTrue/assertIn/

Better error messages.

* test_functions: drop ExpectedException

while this is still unittest and uses assertRaises, same as the rest
of the code, I'm going to use assertRaises, rather than pytest.raises

* Drop testtools

Use unittest.TestCase instead

* s/assertEquals/assertEqual/

assertEquals one is deprecated and throws a warning

* flake

mostly de-denting things after the previous commit

* add a stub for py2.7 compat

* 2.7-compat needs to be done this way

otherwise, the old method name is also used in 3.x, which results
in many warnings

* Use order-ignoring comparisons for several namespace things

Apparently, namespaces return those things in no ordering in
particular. That used to be previously asserted via the
order-ignoring assertItemsEqual, but now that we've changed it
to the order-sensitive assertEqual, it must be set()-ed or similar.

Anyway, I find some hilarity in the fact that we specifically
ordered the expected items, eg `[workflows[1], workflows[0]]`,
AND THEN ignored the order, by way of assertItemsEqual.

Which suggests that perhaps namespaces are fucked. But I'm not sure
that ordering is even part of its contract at all.

* flake