RD-3229 When evaluating functions, also allow changing the context
get_property must be allowed to change the context for the evaluation,
because when we hit a `SELF` in the path, now we're potentially running
in the context of another node!

So, for each function, not only update the currently-evaluated-value,
but also update the current context to what the function thinks it
should be.
And get_property thinks the context should be whatever node we actually
got.

The test & its docstring shows an example.