RD-3229 allow changing context when evaluating functions
This ports #890 to 5.2.7

* RD-3229 When evaluating functions, also allow changing the context

get_property must be allowed to change the context for the evaluation,
because when we hit a `SELF` in the path, now we're potentially running
in the context of another node!

So, for each function, not only update the currently-evaluated-value,
but also update the current context to what the function thinks it
should be.
And get_property thinks the context should be whatever node we actually
got.

The test & its docstring shows an example.

* same for get_attribute

switch SELF if it changed (for get_attr, the context is a dict that
contains self/source/target, and self can well change)