Several workflow_context.py refactors (#896)
* WorkflowContext: make NodesAndInstances be composed

Composition over inheritance!

Instead of subclassing that, just make it a component here. This means
that we now _can_ use just super() and the whole hierarchy is way
simpler.
We have to pass through all those methods, but there's not that many
of them.

* wctx.handler: only create a rest_client on-demand

it's not created in `__init__` now, and so all the other inits
(in super calls) don't need to be under a ctx.push anymore

* Refactor WorkflowNodesAndInstancesContainer methods

DRY up the duplicated parts. Factor out common methods for loading
stuff in.

Thanks to that, no need to have yet another place (inside
CloudifyWorkflowContext) fetching the nodes/instances. Instead,
we can just call refresh, and that is now the only thing that
needs a restclient (ie. needs ctx)

* Move the TaskDispatcher parts to a separate module

workflow_context.py is waaay too big as it is, but I can factor those
parts out to a separate module easily
This is just moving code around, not actually doing any changes to it.
Feel free to not read it too deeply, then.

* CloudifyWorkflowContextInternal: create tasks graph on demand

only create it on first use. That means we don't need workflow
context for the `__init__`.

* CloudifyWorkflowContextInternal: make bootstrap_context into a property

a _get method is nonsense, just make it a property. All that's needed
is a rename pretty much. Look how much nicer!