RD-3415 Add the node instance ID to log messages (#904)
* dispatch.py: remove unused process_registry

..and the other thing too.

I want to remove this first before I go ahead with the other
things, so that this doesn't get in the way.

* Move setup_logging outside of the class

No need to keep it there. It's only used at top-level anyway.

* Remove futile defensive coding

In the `except` branch, we try to access attributes on `handler`,
so if it's None, then we error out anyway. So there's no point
trying to defend against it.

* Remove duplicated except branch

It's identical to the BaseException case, which covers this one
as well obviously.

* remove unused attribute

* Add a way to skip sending a log

We want to emit the log to the logger & logfile, but avoid sending
it to the manager. So the "skip" logic needs to be deeper.

I don't love this, but I don't see a better way currently. :(

* Prefix logfile messages with the node-instance ID

More useful this way!

* kwargs the mocks