Expose .update_node_instance on workflow_context
It's going to be useful to have a wctx.update_ni(), because otherwise,
for all the update calls we want to do (which we currently do in
some places, eg. dep-update), we'd have to use the rest-client
directly.

Using the rest-client directly means no compat with cfy local.

The way to achieve compat with cfy local, is to always use the wctx
as a facade for all the api calls. And so, let's expose this too.