RD-3675 Add the check_status workflow (#955)
* RD-3673 Add system_properties to ni.update()

This is going to be used internally, to set systeh-properties, from
the mgmtworker. There's nothing stopping a user from using it as well,
if they truly want to break their stuff.

* Expose .update_node_instance on workflow_context

It's going to be useful to have a wctx.update_ni(), because otherwise,
for all the update calls we want to do (which we currently do in
some places, eg. dep-update), we'd have to use the rest-client
directly.

Using the rest-client directly means no compat with cfy local.

The way to achieve compat with cfy local, is to always use the wctx
as a facade for all the api calls. And so, let's expose this too.

* local workflows storage: add system_properties and force for NIs

Similar to how we've added system_properties to the rest-client,
we can also add it to the other backend, which is the cfy local
backend, which is this Storage class.

Also add other missing things in there, ie. force and relationships

* Implement the check_status workflow

It's ALMOST a copy of execute_operation, but it sets success and
failure handlers on each task.
Those handlers will know which instance their task ran on, based on
task.info.

* Add a test

* Prettify WorkflowFailed error output

When there's no error text, avoid showing `WorkflowFailed: -> `.
Instead, just show `WorkflowFailed`.

* flake