untangle env/storage
I really don't like the idea of Environment taking the storage
as an argument, and then also loading on the storage, and then
also doing parsing... which object "owns" which object, then?

How about this instead, an Environment is now created from a loaded
storage. In principle, multiple Environments could be loaded from
a single storage. (not yet, but soon)