RD-3900 Add the concept of deployments to local workflows (#963)
* untangle env/storage

I really don't like the idea of Environment taking the storage
as an argument, and then also loading on the storage, and then
also doing parsing... which object "owns" which object, then?

How about this instead, an Environment is now created from a loaded
storage. In principle, multiple Environments could be loaded from
a single storage. (not yet, but soon)

* Add ABC to compat

It's different between py2/py3, and I want to use it.

* replace a relative import with an absolute one

no real reason, I've just been looking at this file quite a few times,
and it pissed me off. Especially with the future import.

* Implement deployments for local workflows

Look, I'm sorry. I'm not gonna split this into multiple commits.
It's going to be impossible to really review anyway.
Passes tests so there's that. (except for a couple, see next commit)

However, here's a few points about this:

- The main idea is adding deployments. This is done by way of having
  _Environment be backed by a DeploymentStorage rather than the base
  Storage (base Storage = FileStorage/InMemoryStorage).
- DeploymentStorage implements all the methods that _Environment uses,
  in practice, (almost) all the methods that were exposed on the base
  Storage before.
- A DeploymentStorage is created by calling Storage.load(dep_id)
  and Storage.create_deployment(..)
- This also means you can have many DeploymentStorages loaded out of
  a single Storage, and each DeploymentStorage can back a single
  _Environment
- For actually storing things, the DeploymentStorage will call to
  its underlying Storage, that it was loaded out of
- I wanted to get rid of InMemoryStorage, but all the tests use it,
  and replacing it with a FileStorage and making sure it was cleaned
  between tests, was too much for right now

* fixup tests

well, those tests asserted on the actual directory structure, and
that has indeed changed