RD-3859 Add workflow_ctx interface for dep-update related calls (#968)
Those are all the calls that are used by dep-update. Currently, the
workflow uses the rest-client directly for them. But I want it to
use the workflow-context, because then, we can also (later) implement
a local-handler for it, and test/use dep-update in local workflows
as well.

This PR is mostly very boring, because all it does, is exposes all
the missing stuf on workflow_ctx.