RD-3941 Handle amqp ca cert given as the cert contents (#970)
* Refactor AMQPParams and broker_ssl_options

This is equivalent but a bit simpler

* RD-3941 AMQPParams: allow passing in the CA as a string

Dont require it in a file, also allow just a string. Python's ssl lib
already does this, so all we need to do is to make sure we pass it
through where needed.

* Still pass in the cert as a file in the http api client

Well, requests still needs a file. Unfortunate, but this is
many orders of magnitude smaller surface, because the rabbitmq api
client is only used when creating tenants etc.
Also, we'll only store the file for the duration of the request,
not for an extended period of time.

* flake