RD-4003 Reverse the condition in wait_after_fail (#971)
* Tasks graph: factor out ._task_error

This bunch of code deserves its own method

* Reverse the wait_after_fail condition

If current time is after the deadline!
Not: if deadline is after current time - that would be the
before-deadline condition.
Hah.

* On error, stop waiting for subgraphs

If we error out, we only allow some time (up to wait_after_fail) to
allow already-started tasks to finish. We don't start new tasks then,
so subgraphs will most likely never finish, so no point waiting for
them.

* Drop test_wait_after_fail

This test didn't detect the bug, and anyway, this is inherently based
on timing. I'll test it in inte-tests instead.

I did try to write a this-level test for it, but it ended up being
SOOOO over-mocked, that it wasn't very useful.