RD-3856 Implement the check_drift workflow
Similar to check_status, but it also runs relationship operations
as well.
The locking in the node-instance-update is a bit heavy-handed, but
at least it's easy to see that it's correct. Maybe someday we'll
replace it with something better (probably a concurrency control thing
like we have for runtime_properties)