RD-3856 Implement the check_drift workflow (#972)
* wctx.update_node_instance: store the updated data

When we call `.update_node_instance()` actually update the
locally-stored instance as well. This way, multiple calls to
`.update_node_instance()` won't overwrite each other's changes.

* local workflows: make sure to return the stored instance

the remote handler already did this, via the restclient, but the local
handler must now do this as well, now that we actually use the return
in update_node_instance

* Add a test for .on_success

Not really related to the rest of the PR, but it felt to me that it
doesn't work, and we didn't really have much tests for this.

Turns out it does work, but let's keep the test.

* RD-3856 Implement the check_drift workflow

Similar to check_status, but it also runs relationship operations
as well.
The locking in the node-instance-update is a bit heavy-handed, but
at least it's easy to see that it's correct. Maybe someday we'll
replace it with something better (probably a concurrency control thing
like we have for runtime_properties)

* Add a test for the check_drift workflow

Check that the right operations run, and store the results in
system_properties

* flake

* set() the keys in tests

py2.7 compat :(

* return .cont() from the success handler

ignoring a success doesn't mean anything