RD-4096 Update: delegate to the new workflow
Normally, blueprints are going to use the builtin update workflow
(the one in cloudify_system_workflows);  but old (6.2 and earlier)
blueprints still have the update workflow defined in the yaml, and
those point to this function;  so, let's make this function just call
the builtin one.

Also, remove those tests, because this workflow doesn't live here
anymore.