RD-4153 `deployment_id` constraint for `node_template` (#1023)
* Pass a `deployment_id` as a constraint in deployment capabilities client

This is more in line with the general use-case of the
`/searches/capabilities` endpoint.  In general, We are more likely to
pass constraints as a whole, not treating `deployment_id` differently.

* Add `deployment_id` constraint for `node_template`

* Enforce `deployment_id` constraint for certain data types

`capability_value` and `node_template` data types require a
`deployment_id` constraint.

* Don't check if `node_template` exists in the DB

... at least not without a (required) `deployment_id` constriant.

Otherwise `value_getter.get()` would return an error (requesting a
required `deployment_id` constraint).

* Validate values of `node_templates` in the context of other constraints

This way the values are verified with at least one (`deployment_id`)
additional constriant.  So the input/parameter will have to match not
only one of the `valid_values` string, but also be an actually valid
value of one of the node IDs of a specific deployment.