RD-4502 Hide useless tracebacks (incl. DSLParsingExc) (#1025)
* Remove disused .deployments_contexts

It was, a long time ago, used in snap-res. But now it's not. I've had
a mental note to remove it for like a year now, but now I'm going
to change stuff about `.start_local_tasks_processing`, so let's
just remove this, then.

* local_tasks_processing: use a `with` rather than try/finally

* workflow handler: only returh the original traceback

This is a small hack allowing us to only show the user the traceback
of their workflow function, without the several frames added by the
framework itself.

* RD-4502 Skip tracebacks for DSLParsingException

This error is already human-readable and formatted. For showing it
to the user, we can skip the traceback.

The .hide_traceback this is perhaps a bit dirty, but it's going to
be very useful! Soon, we can also consider hiding the tracebacks
of other exception kinds, eg. WorkflowFailed itself.

* flake

* reraise the exception `from None`

...actually using this stupid syntax.
Makes the output much better. Hides the parts where the exception was
in the _WorkflowFuncError state.

* Compat for local workflows

Also the mock needs to be a magicmock because `__enter__`

* Whoops. Better return before .get()ing all these

It was the case before, too. In case of an error, we don't want to
.get() all the tasks.

Now, I do recognize it's a bit unwieldy, not throwing, but returning.
Although it makes the traceback nicer, so oh well.