RD-4623 Non-graph execs: handle errors same as graph execs
Well this needs to go under all the `except`s, same as the regular
func call.

FWIW this is about workflows that create operations but not in the
context of a graph.
As in, yes, it is allowed to do, like
```python
def workflow(ctx):
    instance = ctx.node_instances[0]
    instance.execute_operation('foo.bar')
```
And that operation WILL be awaited-for, in just these lines I'm moving.
So let's handle errors in them.