RD-4164 `node_instances.list()` can query with constraints
The function is going to be used in a similar manner as the `list()`
functions for other Cloudify objects (deployments, nodes etc.).  If it's
called with `constraints` dictionary, then instead of just fetching
node_instances with GET and some parameters, we're going to use the
`/searches/node-instances` endpoint to execute a more parametrized
query.