RD-4921 dispatch: reduce import overhead (#1046)
* move is_agent_alive to amqp_client

This function is rarely used, so let's not put it in the main utils
module - it requires importing amqp_client anyway, which pulls in
pika, which is not small.

Now, utils doesn't need pika, nor does it need cloudify.amqp_client

* Inline uuid.uuid4

I'd like to avoid importing uuid on the critical path (when spawning
an operation subprocess). See the docstring as to why.

* Move a bunch of imports into functions

Rarely-used things don't need to be imported all the time. You only
pay the price if you actually use it.

* dsl_parser._compat - fix a merge fail

* flake

* also dont import jinja2 if there's nothing to render